Add deletion of cursor line after brace completion
This commit adds an implementation of automatically deleting the cursor
line if it is blank after a multi-line brace autocompletion was
performed. The motivation for this addition is convenience: immediately
leaving Insert mode after triggering a multi-line brace autocompletion
(e.g., to delete and put text between the braces) saves deleting the
empty line manually.

The current implementation of this behavior introduces a bug: entering
and leaving Insert mode an impractical and inordinate number of times
will unintentionally trigger this behavior due to integer wraparound.
Fortunately, it is nearly impossible for this bug to surface and, if it
does, the consequence is negligible.