Fix <Tab> bug and quote bug
Previously, entering a <Tab> in Insert mode would always skip a quote,
regardless if it was an opening or closing quote. The problem was that
Autodeli_tab() did not identify whether a quote was opening or closing.
Thus, this commit fixes this bug by implementing such a check.
Additionally, it also fixes a bug in Autocomplete_quotes() resulting
from a negative index getting the last character in a string (Vim9
behavior) instead of returning an empty string (legacy script behavior).
Finally, this commit edits and reformats text in a few places to improve
readability.