Keep <SID> for expression mappings
This commit returns the "<SID>" strings, which were removed in Commit
ecb40f4 (Convert legacy script to Vim9 script, 2022-01-28), for
expression mappings. The reason for this change is so that the
script-local functions can still be located if the expression mappings
are defined in a different script (e.g., a session script made via
":mksession").