Rename a klenlib function and add a function
Since the klen library is not being used beyond this plugin,
interface-breaking changes can still be made to it; the
Byteidx_quote_positions() function name is shortened to
Bidx_quote_positions() so that it is more concise but still readable.
Additionally, a function that converts a screen column to the
corresponding byte index in a string is added.