Extend Autodeli_eat() to support new CTRL-U
Recently, the Insert-mode command CTRL-U was changed so that, along with
its original behavior, the command automatically adjusts the indent of
the current line if the line becomes blank. (The documentation for the
command was updated in Vim version 8.2.4310.) Thus, this commit responds
to this change by extending Autodeli_eat(): an intermediary auxiliary
function is defined to handle CTRL-U uniquely while preserving this
plugin's original implementation for overloading backspacing characters.