Abandon local copy of klenlib
Keeping a local copy of the klen library is awkward to maintain and does
not directly represent the relationship between it and this plugin.
(That is, klenlib is a library, and its application should not be
limited to autodeli.) Thus, this copy is removed; the klen library must
be obtained separately.