Define an interface command for Autodeli
Previously, Autodeli was always active when it was installed. It worked
when Vim began, and it affected all buffers. It, however, could not be
disabled easily while Vim was running. The easiest way to disable it was
outside Vim: the user could set the global variable "g:loaded_autodeli"
in his vimrc file or uninstall the plugin. Of course, this method
incurred the same problem, only inverted: Autodeli could not enabled
within Vim.

To address Autodeli's inflexibility, this commit defines a command
called "Autodeli" that acts as an interface to the plugin. The command
allows the user to enable or disable Autodeli in Vim, specifically for
the current buffer or all buffers in the buffer list.

To retain the plugin's original behavior of always being active and save
the user from monotonously setting Autodeli to a preferred state when
editing a new buffer, new buffers inherit the state of Autodeli from
their previous buffers. That is, opening a new buffer from one with
Autodeli enabled will turn the plugin on for that buffer, and vice
versa.