Add an autocommand that untracks deleted buffers
An autocommand that removes a buffer from the list of buffers Autodeli
tracks is defined. This addition is made because Autodeli would track
new buffers permanently, even when they were deleted later. If a new
buffer was edited, deleted, and then edited again, Autodeli might report
the wrong state (e.g., "on" when Autodeli is actually "off"). Therefore,
this commit corrects this issue by implementing the expected behavior of
untracking deleted buffers.