Revise automatic deletion of blank line
The initial implementation of the automatic deletion of the blank line
created from a multi-line brace autocompletion (Commit 139a285 - Add
deletion of cursor line after brace completion) suffered from a couple
of shortcomings. The first was the integer wraparound noted in the
implementation's commit message. The second was the weak requirements
for performing the deletion. If the cursor was on a blank line after a
multi-line brace autocompletion occurred and then Insert mode was
exited, the line was deleted. This behavior allowed the deletion to
trigger unexpectedly, such as when backspacing immediately after a
multi-line brace autocompletion was performed.

This commit corrects the latter issue by triggering the line deletion if
and only if the cursor was on the blank line that the multi-line brace
autocompletion created before leaving Insert mode. It also fixes the
former problem by calling a function after such an autocompletion is
effected rather than using a counter to check if an autocompletion
occurred.